requirements.txt
google-api-python-client
google-genai>=1.0.0
streamlit==1.28.1
pandas==2.1.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv
requests
